[go]: Get compiler version in a more robust way.
Previously, we were scanning for the Go compiler version in
.rodata. This often works, but is error-prone, since that section
can in general contain anything. For example, I saw some binaries that
had the string "go1.9" embedded in .rodata even though they were not
built with Go 1.9.

After this commit, we first try finding the build ID the same way
`debug/buildinfo` from the Go standard library does it (usually in the
.go.buildinfo section). Most of the code was taken from
[here](https://cs.opensource.google/go/go/+/refs/tags/go1.22.1:src/debug/buildinfo/buildinfo.go;l=101),
with a few changes by me to remove all non-ELF-related code and to use
StreamReader for reading.